+++ Mutagen.Bethesda.Records/Oblivion/Records/Major Records/Worldspace.cs
@@ -513,7 +513,7 @@ namespace Mutagen.Bethesda.Oblivion
                 this._OffsetLengthLocation = stream.Position - offset;
             }
 
-            partial void CustomEnd(IBinaryReadStream stream, long finalPos, int offset)
+            partial void CustomEnd(IBinaryReadStream stream, int finalPos, int offset)
             {
                 if (stream.Complete) return;
                 var groupMeta = this._package.Meta.GetGroup(stream);
